+++ cypress/integration/unit_tests/color_function_util_test.js
@@ -1,4 +1,4 @@
-import {getLinearGradient, populateColorFunctions, withColor} from '../../../docs/import/color_function_util.js';
+import {populateColorFunctions, withColor} from '../../../docs/import/color_function_util.js';
 import * as manageLayersLib from '../../../docs/import/manage_layers_lib.js';
 import {getCurrentLayers} from '../../../docs/import/manage_layers_lib.js';
 import {createTrs, setDisasterAndLayers} from '../../support/import_test_util.js';
@@ -164,38 +164,6 @@ describe('Unit tests for color function utility', () => {
     expect(colorFunctionEditor.is(':visible')).to.be.false;
     expect(writeToFirebaseStub).to.not.be.called;
   });
-});
-
-it('creates the correct linear gradients', () => {
-  const layer1 = {
-    'color-function': {
-      'current-style': 0,
-      'color': 'yellow',
-    },
-  };
-
-  const layer2 = {
-    'color-function': {
-      'current-style': 1,
-      'colors': ['yellow', 'red'],
-    },
-  };
-
-  const layer3 = {
-    'color-function': {
-      'current-style': 2,
-      'color': 'blue',
-    },
-  };
-  const layer1Gradient = 'linear-gradient(to right, white, yellow)';
-  expect(getLinearGradient(layer1['color-function'])).to.equal(layer1Gradient);
-
-  const layer2Gradient =
-      'linear-gradient(to right, yellow 0%, yellow 50%, red 50%, red 100%)';
-  expect(getLinearGradient(layer2['color-function'])).to.equal(layer2Gradient);
-
-  const layer3Gradient = 'linear-gradient(to right, blue, blue)';
-  expect(getLinearGradient(layer3['color-function'])).to.equal(layer3Gradient);
 });
 
 /**
