+++ docs/import/color_function_util.js
@@ -2,7 +2,7 @@ import {colorMap, ColorStyle} from '../firebase_layers.js';
 
 import {getCurrentLayers, getRowIndex, ILLEGAL_STATE_ERR, setStatus, updateLayersInFirestore} from './manage_layers_lib.js';
 
-export {getLinearGradient, populateColorFunctions, withColor};
+export {populateColorFunctions, withColor};
 // Visible for testing.
 export {
   setColor,
@@ -457,36 +457,4 @@ function createColorBox(color) {
 function getColorFunction() {
   const index = getRowIndex(globalTd.parents('tr'));
   return getCurrentLayers()[index]['color-function'];
-}
-
-/**
- * Gets the linear gradient of the colors for the legend.
- *
- * @param {Object} colorFunction color data from the layer
- * @return {string} the linear gradient
- */
-function getLinearGradient(colorFunction) {
-  if (!colorFunction) {
-    return '';
-  }
-  const currentStyle = colorFunction['current-style'];
-  let gradientString = 'linear-gradient(to right';
-  switch (currentStyle) {
-    case 0:
-      gradientString += ', white, ' + colorFunction['color'];
-      break;
-    case 1:
-      const colors = [...(new Set(Object.values(colorFunction['colors'])))];
-      const percent = 100 / colors.length;
-      for (let i = 1; i <= colors.length; i++) {
-        gradientString += ', ' + colors[i - 1] + ' ' + (i * percent - percent) +
-            '%, ' + colors[i - 1] + ' ' + i * percent + '%';
-      }
-      break;
-    case 2:
-      gradientString +=
-          ', ' + colorFunction['color'] + ', ' + colorFunction['color'];
-      break;
-  }
-  return gradientString + ')';
 }
